+++ ontimize/config/o-directives.ts
@@ -2,11 +2,13 @@ import { OTabGroupDirective } from '../directives/o-tab-group.directive';
 import { OKeyboardListenerDirective } from '../directives/keyboard-listener.directive';
 import { OListItemDirective } from '../components/list/list-item/o-list-item.directive';
 import { OLockerDirective } from '../directives/locker.directive';
+import { OHiddenDirective } from '../directives/o-hidden.directive';
 
 export const ONTIMIZE_DIRECTIVES = [
   // Ontimize directives...
   OKeyboardListenerDirective,
   OListItemDirective,
   OTabGroupDirective,
-  OLockerDirective
+  OLockerDirective,
+  OHiddenDirective
 ];