+++ ontimize/config/o-providers.ts
@@ -1,39 +1,40 @@
-import { Injector, Provider } from '@angular/core';
 import { LOCATION_INITIALIZED } from '@angular/common';
+import { Injector, Provider } from '@angular/core';
 import { BaseRequestOptions, XHRBackend } from '@angular/http';
 import { Router } from '@angular/router';
-import { OHttp } from '../util/http/OHttp';
-import { Events } from '../util/events';
-import { AppConfig, Config } from '../config/app-config';
+import { combineLatest } from 'rxjs';
 
+import { OContextMenuService } from '../components/contextmenu/o-context-menu.service';
+import { AppConfig, Config } from '../config/app-config';
 import {
-  LoginService,
-  NavigationService,
-  OntimizeService,
-  OntimizeFileService,
-  OntimizeExportService,
-  MomentService,
-  NumberService,
-  CurrencyService,
-  OTranslateService,
-  DialogService,
-  SnackBarService,
-  AuthGuardService,
-  dataServiceFactory,
-  LocalStorageService,
   appConfigFactory,
   AppMenuService,
-  OUserInfoService,
+  AuthGuardService,
+  CurrencyService,
+  dataServiceFactory,
+  DialogService,
+  LocalStorageService,
+  LoginService,
+  MomentService,
+  NavigationService,
+  NumberService,
   OModulesInfoService,
+  OntimizeExportService,
+  OntimizeFileService,
+  OntimizeMatIconRegistry,
+  OntimizeService,
   OntimizeServiceResponseParser,
-  OntimizeMatIconRegistry
+  OTranslateService,
+  OUserInfoService,
+  SnackBarService
 } from '../services';
-
 import { OFormLayoutManagerService } from '../services/o-form-layout-manager.service';
-import { OContextMenuService } from '../components/contextmenu/o-context-menu.service';
-import { Codes } from '../util/codes';
 import { Error403Component } from '../services/permissions/error403/o-error-403.component';
+import { ORemoteConfigurationService } from '../services/remote-config.service';
 import { ShareCanActivateChildService } from '../services/share-can-activate-child.service';
+import { Codes } from '../util/codes';
+import { Events } from '../util/events';
+import { OHttp } from '../util/http/OHttp';
 
 function addPermissionsRouteGuard(injector: Injector) {
   const route = injector.get(Router);
@@ -45,6 +46,7 @@ function addPermissionsRouteGuard(injector: Injector) {
 
 export function appInitializerFactory(injector: Injector, config: Config, oTranslate: OTranslateService) {
   return () => new Promise<any>((resolve: any) => {
+    let observableArray = [];
     const locationInitialized = injector.get(LOCATION_INITIALIZED, Promise.resolve(null));
     locationInitialized.then(() => {
       oTranslate.setDefaultLang('en');
@@ -64,12 +66,21 @@ export function appInitializerFactory(injector: Injector, config: Config, oTrans
       if (userLang && config.applicationLocales.indexOf(userLang) === -1) {
         config.applicationLocales.push(userLang);
       }
-
-      oTranslate.setAppLang(userLang).subscribe(resolve, resolve, resolve);
+      if (config['uuid'] === undefined || config['uuid'] === null || config['uuid'] === '') {
+        console.error('Your app must have an \'uuid\' property defined on your app.config file. Otherwise, your application will not work correctly.');
+        alert('Your app must have an \'uuid\' property defined on your app.config file. Otherwise, your application will not work correctly.');
+      }
+      injector.get(NavigationService).initialize();
+      injector.get(OntimizeMatIconRegistry).initialize();
+      injector.get(LocalStorageService).setBackwardCompatibility();
+      addPermissionsRouteGuard(injector);
+      observableArray.push(oTranslate.setAppLang(userLang));
+      const remoteConfigService = injector.get(ORemoteConfigurationService);
+      observableArray.push(remoteConfigService.initialize());
+      combineLatest(observableArray).subscribe(() => {
+        resolve();
+      });
     });
-    injector.get(NavigationService).initialize();
-    injector.get(OntimizeMatIconRegistry).initialize();
-    addPermissionsRouteGuard(injector);
   });
 }
 
@@ -163,6 +174,10 @@ export function getOModulesInfoServiceProvider(injector: Injector) {
 
 export function getOntimizeServiceResponseParser(injector: Injector) {
   return new OntimizeServiceResponseParser(injector);
+}
+
+export function getORemoteConfigurationService(injector: Injector): ORemoteConfigurationService {
+  return new ORemoteConfigurationService(injector);
 }
 
 export const ONTIMIZE_PROVIDERS: Provider[] = [
@@ -287,5 +302,10 @@ export const ONTIMIZE_PROVIDERS: Provider[] = [
   {
     provide: ShareCanActivateChildService,
     useClass: ShareCanActivateChildService
+  },
+  {
+    provide: ORemoteConfigurationService,
+    useFactory: getORemoteConfigurationService,
+    deps: [Injector]
   }
 ];