+++ ontimize/util/util.ts
@@ -242,9 +242,13 @@ export class Util {
    * Returns the provided string in lowercase and without accent marks.
    * @param value the text to normalize
    */
-  static normalizeString(value: string): string {
+  static normalizeString(value: string, toLowerCase: boolean = true): string {
     if (value && value.length) {
-      return value.normalize('NFD').replace(/[\u0300-\u036f]/g, '').toLowerCase();
+      let result = value.normalize('NFD').replace(/[\u0300-\u036f]/g, '');
+      if (toLowerCase) {
+        result = result.toLowerCase();
+      }
+      return result;
     }
     return '';
   }